Indicate test result in stdout instead of exit code
Jenkins seems to check the exit code of all individual commands in the
execute shell. As soon as Jenkins detects a non-zero exit code for
any command in the execute shell, it would skip the remaining commands
and flag the build as overall failure right away. Sometimes this is
not what we want, e.g. if there are cleanup steps to be done after the
affected command. To work around this behavior, we ensure that the
smoke test always terminates with exit code zero, and indicate the
pass/fail result in stdout.