Handle certain degenerate cases in polygon intersection
When a vertex of polygon A was located too close to an edge in polygon B,
the main polygon intersection algorithm would sometimes behave
incorrectly. This fix attempts to work around this by perturbing the
edge-touching vertices in such cases, hopefully leading the main
algorithm to detect intersections in the expected way.