Added 'WI keyword implicit' mode.
In 'WI keyword implicit' mode, the expression is considered to be a
sequence of lon/lat coordinates, but the 'WI' keyword must not appear
in front (it would be considered a syntax error). 'WI keyword implicit'
mode is relevant only in 'WI only' mode, i.e. when only a WI expression
is allowed.